+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * SmartSelfie Enrollment and Authentication receive new, streamlined UIs
   * Set `useStrictMode=true` parameter for better pass rates
   * Removed `showInstructions` and `jobId` input parameters
+* Fixed a bug where some failed authentication requests were incorrectly handled
+* Fixed a bug where errors with no code were not being handled correctly
 
 ## 10.1.6