+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Unreleased
+## 10.2.0
 
 * Added an optional "Strict Mode" to SmartSelfie Enrollment and Authentication to achieve better pass rates. Set `useStrictMode=true` to enable this new, streamlined UI and associated active liveness tasks
 * Fixed a bug where some failed authentication requests were incorrectly handled